Move Dexsuite procedural cuboids out of the shared object library
Keep Newton-only table/cube assets local to the Dexsuite lift example and
construct them directly instead of registering them globally.
